[ECS] Refactor CheckCommand
Move all print functionality inside CheckCommandReporter
